Safari-609.1.11.1/LayoutTests/accessibility/gtk/text-for-range-extraneous-whitespace.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
</head>
<body id="body">
<p>This
                          paragraph
                                                      is
                                                                                                                                                                                                                                                                                                                                                                            borked!</p>
<p id="description"></p>
<div id="console"></div>
<script>
    description("This tests the ability to get element text for a specified range.");
    if (window.accessibilityController) {
        document.getElementById("body").focus();
        var webArea = accessibilityController.focusedElement;
        var paragraph = webArea.childAtIndex(0);
        shouldBe("paragraph.stringForRange(0, 25)", "'This paragraph is borked!'");
    }
</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
